I need to learn to read the docs more carefully.
Changed the way the custom resource handler works. Instead of using the crhelper, the proper way is to use the skeleton lambda in the docs to generate your lambda handler instead. Connect needs the is_complete handler because the instance status takes about a minute to become active. This means that any calls, like creating a user fails after.